feat(bottle): add server API for my
